+++ src/main/java/oms/pc_protector/restApi/statistics/model/ResponseVO.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ResponseVO {
 
-    private int departmentCode;
+    private Long departmentCode;
     private String yearMonth;
