+++ src/main/java/oms/pc_protector/restApi/user/model/UserSearchInputVO.java
@@ -10,6 +10,6 @@ public class UserSearchInputVO {
     private String userId;
     private String name;
     private String phone;
-    private Integer departmentCode;
+    private Long departmentCode;
 
 }
